Se ha corregido errores y se han terminado todos los service

+++ src/main/java/com/newnocturnalhunter/api_rest/repository/PersonajesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonajesRepository extends JpaRepository<Personajes, Integer> {
+public interface PersonajesRepository extends JpaRepository<Personajes, Long> {
 }
